still working on leave duration and date picker

+++ database/migrations/2022_08_29_132035_create_employees_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('last_name');
             $table->string('email')->unique();
             $table->string('phone_number')->unique();
-            $table->integer('nb_of_days')->default(15);
+            $table->double('nb_of_days')->default(15);
             $table->foreignId('department_id')->nullable()->constrained()->onUpdate('CASCADE')->onDelete('SET NULL');
             $table->timestamps();
             $table->softDeletes();
